// backend-nestjs/src/common/menu-types.ts
export enum MenuLocation {
  HEADER = 'HEADER',
  FOOTER = 'FOOTER',
  SIDEBAR = 'SIDEBAR',
  MOBILE = 'MOBILE',
}

export enum MenuStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
}

export enum MenuItemType {
  PAGE = 'PAGE',
  ARTICLE = 'ARTICLE',
  POST = 'POST',
  GALLERY = 'GALLERY',
  TENANT_ROUTE = 'TENANT_ROUTE',
  CUSTOM_URL = 'CUSTOM_URL',
  EXTERNAL_URL = 'EXTERNAL_URL',
  GROUP = 'GROUP',
}

export enum MenuItemTarget {
  SELF = 'SELF',
  BLANK = 'BLANK',
}